move json out of inner classes
