+++ ethereum/api/src/main/java/org/hyperledger/besu/ethereum/api/jsonrpc/internal/parameters/JsonCallParameter.java
@@ -19,8 +19,8 @@ import static java.lang.Boolean.FALSE;
 import org.hyperledger.besu.ethereum.core.Address;
 import org.hyperledger.besu.ethereum.core.Gas;
 import org.hyperledger.besu.ethereum.core.Wei;
-import org.hyperledger.besu.ethereum.core.deserializer.GasDeserializer;
-import org.hyperledger.besu.ethereum.core.deserializer.HexStringDeserializer;
+import org.hyperledger.besu.ethereum.core.json.GasDeserializer;
+import org.hyperledger.besu.ethereum.core.json.HexStringDeserializer;
 import org.hyperledger.besu.ethereum.transaction.CallParameter;
 
 import java.util.Optional;
