+++ ethereum/core/src/main/java/org/hyperledger/besu/ethereum/core/AccessListEntry.java
@@ -16,23 +16,16 @@ package org.hyperledger.besu.ethereum.core;
 
 import static com.google.common.base.Preconditions.checkState;
 
-import java.io.IOException;
-import java.util.ArrayList;
 import java.util.List;
 
-import com.fasterxml.jackson.core.JsonGenerator;
-import com.fasterxml.jackson.core.JsonParser;
-import com.fasterxml.jackson.core.JsonToken;
-import com.fasterxml.jackson.databind.DeserializationContext;
-import com.fasterxml.jackson.databind.SerializerProvider;
 import com.fasterxml.jackson.databind.annotation.JsonDeserialize;
 import com.fasterxml.jackson.databind.annotation.JsonSerialize;
-import com.fasterxml.jackson.databind.deser.std.StdDeserializer;
-import com.fasterxml.jackson.databind.ser.std.StdSerializer;
 import org.apache.tuweni.bytes.Bytes32;
+import org.hyperledger.besu.ethereum.core.json.AccessListEntryDeserializer;
+import org.hyperledger.besu.ethereum.core.json.AccessListEntrySerializer;
 
-@JsonSerialize(using = AccessListEntry.Serializer.class)
-@JsonDeserialize(using = AccessListEntry.Deserializer.class)
+@JsonSerialize(using = AccessListEntrySerializer.class)
+@JsonDeserialize(using = AccessListEntryDeserializer.class)
 public class AccessListEntry {
   private final Address address;
   private final List<Bytes32> storageKeys;
@@ -51,57 +44,4 @@ public class AccessListEntry {
     return storageKeys;
   }
 
-  public static class Serializer extends StdSerializer<AccessListEntry> {
-
-    Serializer() {
-      this(null);
-    }
-
-    protected Serializer(final Class<AccessListEntry> t) {
-      super(t);
-    }
-
-    @Override
-    public void serialize(
-        final AccessListEntry accessListEntry,
-        final JsonGenerator gen,
-        final SerializerProvider provider)
-        throws IOException {
-      gen.writeStartObject();
-      gen.writeFieldName("address");
-      gen.writeString(accessListEntry.getAddress().toHexString());
-      gen.writeFieldName("storageKeys");
-      final List<Bytes32> storageKeys = accessListEntry.getStorageKeys();
-      gen.writeArray(
-          storageKeys.stream().map(Bytes32::toHexString).toArray(String[]::new),
-          0,
-          storageKeys.size());
-      gen.writeEndObject();
-    }
-  }
-
-  public static class Deserializer extends StdDeserializer<AccessListEntry> {
-    private Deserializer() {
-      this(null);
-    }
-
-    protected Deserializer(final Class<?> vc) {
-      super(vc);
-    }
-
-    @Override
-    public AccessListEntry deserialize(final JsonParser p, final DeserializationContext ctxt)
-        throws IOException {
-      checkState(p.nextFieldName().equals("address"));
-      final Address address = Address.fromHexString(p.nextTextValue());
-      checkState(p.nextFieldName().equals("storageKeys"));
-      checkState(p.nextToken().equals(JsonToken.START_ARRAY));
-      final ArrayList<Bytes32> storageKeys = new ArrayList<>();
-      while (!p.nextToken().equals(JsonToken.END_ARRAY)) {
-        storageKeys.add(Bytes32.fromHexString(p.getText()));
-      }
-      p.nextToken(); // consume end of object
-      return new AccessListEntry(address, storageKeys);
-    }
-  }
 }
