+++ ethereum/core/src/main/java/org/hyperledger/besu/ethereum/core/json/GasDeserializer.java
@@ -12,27 +12,28 @@
  *
  * SPDX-License-Identifier: Apache-2.0
  */
-package org.hyperledger.besu.ethereum.core.deserializer;
+package org.hyperledger.besu.ethereum.core.json;
+
+import org.hyperledger.besu.ethereum.core.Gas;
 
 import java.io.IOException;
 
 import com.fasterxml.jackson.core.JsonParser;
 import com.fasterxml.jackson.databind.DeserializationContext;
 import com.fasterxml.jackson.databind.deser.std.StdDeserializer;
-import org.apache.tuweni.bytes.Bytes;
 
-public class HexStringDeserializer extends StdDeserializer<Bytes> {
-  public HexStringDeserializer() {
+public class GasDeserializer extends StdDeserializer<Gas> {
+  public GasDeserializer() {
     this(null);
   }
 
-  public HexStringDeserializer(final Class<?> vc) {
+  public GasDeserializer(final Class<?> vc) {
     super(vc);
   }
 
   @Override
-  public Bytes deserialize(final JsonParser jsonparser, final DeserializationContext context)
+  public Gas deserialize(final JsonParser jsonparser, final DeserializationContext context)
       throws IOException {
-    return Bytes.fromHexString(jsonparser.getCodec().readValue(jsonparser, String.class));
+    return Gas.fromHexString(jsonparser.getCodec().readValue(jsonparser, String.class));
   }
 }
